pyproject.toml
[tool.poetry]
name = "stock-crawler"
version = "0.1.0"
description = ""
authors = ["LittleWhiteYA <f46284628@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.25.1"
beautifulsoup4 = "^4.9.3"
pymongo = "^3.11.3"
python-dotenv = "^0.17.0"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
